Add Gradle wrapper (5.1.1)
This adds a common build infrastructure to this project.
I have used gradle 5.1.1, because I see no reason to stay
with 4.7 everywhere. We are going to update anyway.

We could use Maven, but I have more experiences with
Gradle. Moreover, typical user will not build this
plugin on his own.